Add files via upload
The fit type was hard-coded as 'moffat' under the parallelized fitting option, this has been corrected to the 'fit_type' variable set by the user. When 'dual_psf' was set to False the output filename had a trailing '_' before the '.fits' extension, this has been corrected in the "Reinterpolate Stack" cell.